+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/shared/entity-section-empty-state.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from "@mui/material";
 import { ReactNode } from "react";
-import { WhiteCard } from "../../entity-types/white-card";
+import { WhiteCard } from "../../../entity-types/white-card";
 
 interface EntitySectionEmptyStateProps {
   title: string;
