+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/shared/entity-section.tsx
